Refactor Cpu and Program structures; update instruction handling and add debug output

+++ src/program/mod.rs
@@ -1,9 +1,10 @@
 use crate::instruction::Instruction;
 use crate::instruction::exec_instruction;
 
+#[derive(Debug)]
 pub struct Program {
     pub instructions: Vec<Instruction>,
-    pub idx_instr: u32,
+    pub idx_instr: usize,
     pub memory: [i32; 32],
 }
